refactor: move osinstall into osctl
