+++ .conform.yaml
@@ -25,7 +25,6 @@ policies:
       - kernel
       - osctl
       - osd
-      - osinstall
       - rootfs
       - tools
       - trustd
