+++ cmd/osctl/cmd/root.go
@@ -27,6 +27,7 @@ var (
 	kubernetes   bool
 	talosconfig  string
 	target       string
+	userdataFile string
 )
 
 // rootCmd represents the base command when called without any subcommands
