--- cmd/osinstall/cmd/version.go
@@ -1,39 +1,0 @@
-/* This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
-
-package cmd
-
-import (
-	"fmt"
-	"os"
-
-	"github.com/spf13/cobra"
-	"github.com/talos-systems/talos/internal/pkg/version"
-)
-
-var (
-	shortVersion bool
-)
-
-// versionCmd represents the version command
-var versionCmd = &cobra.Command{
-	Use:   "version",
-	Short: "Prints the version",
-	Long:  ``,
-	Run: func(cmd *cobra.Command, args []string) {
-		if shortVersion {
-			version.PrintShortVersion()
-		} else {
-			if err := version.PrintLongVersion(); err != nil {
-				fmt.Println(err)
-				os.Exit(1)
-			}
-		}
-	},
-}
-
-func init() {
-	versionCmd.Flags().BoolVar(&shortVersion, "short", false, "Print the short version")
-	rootCmd.AddCommand(versionCmd)
-}
